==> src/network_scripts/PANDA_input_construction/01_Mouse_TFmotif.Rmd <==
---
title: "Mouse_PANDA_TFmotif_01"
author: "Jordan Whitlock"
date: '2022-09-06'
output: html_document
---

# Generating own CIS-BP Tf-Motif input:

* Grabbing TF-Motif Input for Mus Musculus (Mouse):
  * Accessed 221025 from Kimberly Glass Resources ["PANDA-ready Motif Mappings"](https://sites.google.com/a/channing.harvard.edu/kimberlyglass/tools/resources)
  * The resource contains a Mouse [Motif Scan](https://drive.google.com/file/d/12CROLzrirDOAgx5hlUTl-Vhu_LGsOhuR/view?usp=sharing) (file name: Mus_musculus_motifinfo.txt) for mm10 here that contains the TF and its MOTIF ID
  * Additionally a [Prior Network](https://drive.google.com/file/d/1F79g9dh9g3jINsxUpR5qfRwvzFsc-vRW/view?usp=sharing) (file name: mm10_reseq_-750_250_1e-4.txt) is provided that contains MOTIF IDs and the gene that they bind to
  * How motif mapping was generated by [Glass](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0064832#s2): The mappings were produced by using FIMO to scan a genome for sets of  CIS-BP Single Species DNA-motifs curated in the MEME suite. Hits that met a significance less than 1e-4 were recorded and mapped to genes based on annotations downloaded from UCSC. Significant motif hits that fell within [-750,+250] of a gene's transcriptional start site were used to construct a prior regulatory network. 

While mapping motifs to genes, there are three genes that do not have a gene symbol in the input data:
* ENSMUSG00000090020
* ENSMUSG00000079994
* ENSMUSG00000044690


Note: The TF Motif input is needed in order to construct string PPI inputs for the network analysis. This script precedes the 02_Mouse_TF_motif_enrichment.Rmd and 03_Mouse_PANDA_ppi.Rmd. The output generated here is needed as an input for these scripts analysis. 

## Code: 

Load in libraries needed:
```{r}
library(stringr)
library(dplyr)
library(here)
library(styler)
library(lintr)
set.seed(2178)
```


Programmatically read the .txt files into R (see notes and script in accompanying dry lab notebook referenced above). Need to map the MOTIF IDs in the prior network and their corresponding genes to the MOTIF IDs in the Mouse Motif Scan to create a final input data set that contains the following columns: "gene", "TF", "1/0" where the 1 indicates the relationship being present (final input data set should be all 1's)

```{r}
# load in the data from Glass Resource:
mouse_motif <- read.table(here("data/unprocessed/Mus_musculus_motifinfo.txt"),
  header = TRUE
) # has column headers
prior_network <- read.table(here("data/unprocessed/mm10_reseq_-750_250_1e-4.txt")) # does not have column headers
```

### Clean up mouse_motif input data:
```{r}
# remove the version numbers on Motif IDs from MotifID column:
MotifIDs <- str_split(mouse_motif$MotifID, "\\_", simplify = TRUE)[, 1] # grab list of IDs without version
mouse_motif$MotifID <- MotifIDs # set list of cleaned up IDs and MotifID column values

# remove parentheses and other characters from TF info so that TF only contains name of TF:
TFs <- str_split(mouse_motif$TF_Info, "\\_", simplify = TRUE)[, 1] # splits every column value out by separator and only grabs the gene names
TFs <- gsub("[()]", "", TFs) # removes genes from ()
mouse_motif$TF_Info <- TFs
dim(mouse_motif) # 718 total TFs, 2 columns 1 for TF_Info and 1 for MotifID. NOTE: there are three unknown genes that do not have a gene symbol (ENSMUSG00000090020, ENSMUSG00000079994, ENSMUSG00000044690), may need to remove later if cause issues in network construction
```

### Add column names to prior network for mapping:
```{r}
colnames(prior_network) <- c("MotifID", "Motif", "interaction")
head(prior_network)
```


### Map motifs to TF using Motif IDs using innerjoin:
```{r}
mm10_TFmotifs <- inner_join(mouse_motif,
  prior_network,
  by = "MotifID"
) # generate final input data for PANDA

mm10_TFmotifs <- subset(mm10_TFmotifs, select = -c(MotifID)) # drop interaction information and Motif IDs, because all interactions are present (represented as a 1) and this information is not needed for PANDA input
write.table(mm10_TFmotifs,
  file = here("data/processed/motif_inputs/mm10_TFmotifs.txt"),
  sep = "\t"
) # save final input data to Cheaha for use in network construction
```

```{r}
proc.time()
```
   user  system elapsed 
 13.436   2.078 186.854

```{r}
# run style
style_file(here("src/network_scripts/input_construction/Setbp1_PANDA_TFmotif.Rmd"))
# lintr was run as well
```

```{r}
sessionInfo()
```
R version 4.1.3 (2022-03-10)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 20.04.5 LTS

Matrix products: default
BLAS/LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.8.so

locale:
[1] C

attached base packages:
[1] stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
[1] lintr_3.0.2   styler_1.9.0  here_1.0.1    dplyr_1.1.0   stringr_1.5.0

loaded via a namespace (and not attached):
 [1] pillar_1.8.1      compiler_4.1.3    R.methodsS3_1.8.2 R.utils_2.12.2    remotes_2.4.2     tools_4.1.3       digest_0.6.31     evaluate_0.20    
 [9] lifecycle_1.0.3   tibble_3.1.8      R.cache_0.16.0    pkgconfig_2.0.3   rlang_1.0.6       rex_1.2.1         cli_3.6.0         rstudioapi_0.14  
[17] yaml_2.3.7        xfun_0.37         fastmap_1.1.1     cyclocomp_1.1.0   xml2_1.3.3        withr_2.5.0       knitr_1.42        generics_0.1.3   
[25] desc_1.4.2        vctrs_0.5.2       rprojroot_2.0.3   tidyselect_1.2.0  glue_1.6.2        R6_2.5.1          processx_3.8.0    fansi_1.0.4      
[33] rmarkdown_2.20    purrr_1.0.1       callr_3.7.3       magrittr_2.0.3    htmltools_0.5.4   ps_1.7.2          utf8_1.2.3        stringi_1.7.12   
[41] lazyeval_0.2.2    crayon_1.5.2      R.oo_1.25.0
